first-action
Create a new branch. Branches should be named intentionally, so a good name for this branch could be first-action.
On the new branch, create a directory: action-a. Note: If you're working on GitHub.com, you can create a directory and a file at the same time by naming the file action-a/Dockerfile.
In the action-a directory, create a file titled Dockerfile.
Fill the Dockerfile with the content below:
FROM debian:9.5-slim

ADD entrypoint.sh /entrypoint.sh
RUN chmod +x /entrypoint.sh
ENTRYPOINT ["/entrypoint.sh"]